remove custom monads, add ts-results

+++ resources/user/user-service.ts
@@ -1,27 +1,64 @@
+import { Err, ErrImpl, Ok, OkImpl, Result } from "ts-results";
 import { UserSchema } from "./user-model.ts";
 
-async function createUserService(user: UserSchema) {
-    const denoDB = await Deno.openKv("user-db");
-    const result = await denoDB.set(["users", user.id], user);
-    denoDB.close();
-    return result;
+async function createUserService(
+    user: UserSchema,
+): Promise<ErrImpl<string> | OkImpl<UserSchema>> {
+    try {
+        const denoDB = await Deno.openKv("user-db");
+        if (denoDB === null || denoDB === undefined) {
+            return new Err("Error opening database");
+        }
+
+        const result = await denoDB.set(["users", user.id], user);
+        denoDB.close();
+
+        return result.ok ? new Ok(user) : new Err("Error creating user");
+    } catch (error) {
+        return new Err(`Error creating user: ${error ?? "Unknown error"}`);
+    }
 }
 
-async function getUserService(id: string) {
-    const denoDB = await Deno.openKv("user-db");
-    const result = await denoDB.get(["users", id]);
-    denoDB.close();
-    return result;
+async function getUserService(
+    id: string,
+): Promise<ErrImpl<string> | OkImpl<UserSchema>> {
+    try {
+        const denoDB = await Deno.openKv("user-db");
+        if (denoDB === null || denoDB === undefined) {
+            return new Err("Error opening database");
+        }
+
+        const result = await denoDB.get(["users", id]);
+        denoDB.close();
+
+        return result.value !== null
+            ? new Ok(result.value as UserSchema)
+            : new Err("User not found");
+    } catch (error) {
+        return new Err(
+            `Error getting user: ${error ?? "Unknown error"}`,
+        );
+    }
 }
 
 async function getUsersService() {
-    const denoDB = await Deno.openKv("user-db");
-    const iter = denoDB.list({ prefix: ["users"] });
-    const result = [];
-    for await (const { key, value } of iter) {
-        result.push({ key: key[1], value: value });
+    try {
+        const denoDB = await Deno.openKv("user-db");
+        if (denoDB === null || denoDB === undefined) {
+            return new Err("Error opening database");
+        }
+
+        const iter = denoDB.list({ prefix: ["users"] });
+        const result = [];
+        for await (const { key, value } of iter) {
+            result.push({ key: key[1], value: value });
+        }
+        return result;
+    } catch (error) {
+        return new Err(
+            `Error getting users: ${error ?? "Unknown error"}`,
+        );
     }
-    return result;
 }
 
 async function updateUserService(user: UserSchema) {
@@ -30,8 +67,21 @@ async function updateUserService(user: UserSchema) {
 }
 
 async function deleteUserService(id: string) {
-    const denoDB = await Deno.openKv("user-db");
-    return await denoDB.delete(["users", id]);
+    try {
+        const denoDB = await Deno.openKv("user-db");
+        if (denoDB === null || denoDB === undefined) {
+            return new Err("Error opening database");
+        }
+
+        await denoDB.delete(["users", id]);
+        denoDB.close();
+
+        return new Ok("User deleted");
+    } catch (error) {
+        return new Err(
+            `Error deleting user: ${error ?? "Unknown error"}`,
+        );
+    }
 }
 
 export {
